Create README.md
I’m interested in Redhat Linux for free, because Centos, really so sad, will at the end of this year no more supported. Oracle Linux could be an alternative, but the original I think so is much better. I`m retired and can`t afford expensive subscription. I`m not a coder I`m just a user.
Redhat gave the promise, that for single users and developers Redhat linux will be free of charge. So I take Redhat by the word and hope they will stay reliable.
Just wait and see. Hope that by Github account I`m eligible to download the RedhatOS and maintain it with periodic updates.
At the moment I keep going on with Centos8, but will be prepared for Redhat, the original stable and reliable operating linux system. Hope there will be how to instruction for this item. By hpl